switching tdat and dat
