+++ tests/log/coco/test_coco_events.py
@@ -5,7 +5,7 @@ from dataclasses import FrozenInstanceError
 import random
 
 coco_start_params = {
-    "fopt": 0.1,
+    "fopt": -0.1,  # so it is less than mf in eval_params
     "algo": "test_algo",
     "problem": 1,
     "dim": 10,
@@ -33,12 +33,12 @@ end_params = {}
 def test_coco_start():
     event = COCOStart(**coco_start_params)
     assert isinstance(event, LogEvent)
-    assert event.fopt == 0.1
-    assert event.algo == "test_algo"
-    assert event.problem == 1
-    assert event.dim == 10
-    assert event.inst == 1
-    assert event.suite == "test_suite"
+    assert event.fopt == coco_start_params["fopt"]
+    assert event.algo == coco_start_params["algo"]
+    assert event.problem == coco_start_params["problem"]
+    assert event.dim == coco_start_params["dim"]
+    assert event.inst == coco_start_params["inst"]
+    assert event.suite == coco_start_params["suite"]
     assert isinstance(event.exp_id, str)
 
     with pytest.raises(FrozenInstanceError):
@@ -57,8 +57,8 @@ def test_coco_start_custom_exp_id():
 def test_coco_eval():
     event = COCOEval(**eval_params)
     assert isinstance(event, LogEvent)
-    assert event.x == [1.0, 2.0, 3.0]
-    assert event.mf == 0.5
+    assert event.x == eval_params["x"]
+    assert event.mf == eval_params["mf"]
 
 
 def test_coco_end():
