+++ ttex/log/coco/coco_splitter.py
@@ -5,10 +5,10 @@ from ttex.log.formatter import StrRecord
 from ttex.log.coco.record import (
     COCOInfoHeader,
     COCOInfoRecord,
-    COCOdatHeader,
-    COCOdatRecord,
     COCOtdatHeader,
     COCOtdatRecord,
+    COCOdatHeader,
+    COCOdatRecord,
 )
 from cocopp.testbedsettings import SuiteClass
 from typing import List, Dict, Optional
@@ -47,37 +47,37 @@ class COCOKeySplitter(KeySplitter):
         if isinstance(event, COCOStart):
             self._reset_triggers(event.suite)
             info_header = COCOInfoHeader(state)
-            log_dat_header = COCOdatHeader(state)
             log_tdat_header = COCOtdatHeader(state)
-            state.set_dat_filepath(log_dat_header.filepath, info_header.filepath)
+            log_dat_header = COCOdatHeader(state)
+            state.set_dat_filepath(log_tdat_header.filepath, info_header.filepath)
             if info_header.emit():
                 return_dict["info"] = info_header
-            if log_dat_header.emit():
-                return_dict["log_dat"] = log_dat_header
             if log_tdat_header.emit():
                 return_dict["log_tdat"] = log_tdat_header
+            if log_dat_header.emit():
+                return_dict["log_dat"] = log_dat_header
         elif isinstance(event, COCOEval):
-            log_dat_record = COCOdatRecord(state)
             log_tdat_record = COCOtdatRecord(state)
-            if log_dat_record.emit(
+            log_dat_record = COCOdatRecord(state)
+            if log_tdat_record.emit(
                 base_evaluation_triggers=self.base_evaluation_triggers,
                 number_evaluation_triggers=self.number_evaluation_triggers,
-                last_dat_emit=None,
+                last_tdat_emit=None,
             ):
                 # explicitly not the last eval
-                return_dict["log_dat"] = log_dat_record
-                state.last_dat_emit = state.f_evals
-            if log_tdat_record.emit(self.trigger_targets):
                 return_dict["log_tdat"] = log_tdat_record
+                state.last_tdat_emit = state.f_evals
+            if log_dat_record.emit(self.trigger_targets):
+                return_dict["log_dat"] = log_dat_record
         elif isinstance(event, COCOEnd):
             # Emit last evaluation if not already done
-            log_dat_record = COCOdatRecord(state)
-            if log_dat_record.emit(
+            log_tdat_record = COCOtdatRecord(state)
+            if log_tdat_record.emit(
                 base_evaluation_triggers=self.base_evaluation_triggers,
                 number_evaluation_triggers=self.number_evaluation_triggers,
-                last_dat_emit=state.last_dat_emit,
+                last_tdat_emit=state.last_tdat_emit,
             ):
-                return_dict["log_dat"] = log_dat_record
+                return_dict["log_tdat"] = log_tdat_record
             info_record = COCOInfoRecord(state)
             if info_record.emit():
                 return_dict["info"] = info_record
