+++ ttex/log/coco/record/log_tdat.py
@@ -1,20 +1,86 @@
 from ttex.log.coco.record import COCOLogRecord, COCOLogHeader
 from ttex.log.coco import COCOState
 from typing import Optional, List
+import math
 
 
 class COCOtdatRecord(COCOLogRecord):
-    def emit(self, trigger_targets: Optional[List[float]]) -> bool:  # type: ignore[override]
-        if not trigger_targets:
+    @staticmethod
+    def trigger_nth(number_evaluation_triggers: int, f_evals: int) -> bool:
+        """
+        Determine if the current function evaluation count `f_evals` meets the
+        trigger condition based on `number_evaluation_triggers`.
+
+        - every 10**(exponent1/number_of_triggers) for exponent1 >= 0
+        See: https://github.com/numbbo/coco/blob/master/code-experiments/src/coco_observer.c
+        Args:
+            number_evaluation_triggers (int): Number of evaluation triggers.
+            f_evals (int): Current function evaluation count.
+
+        Returns:
+            bool: True if the trigger condition is met, False otherwise.
+        """
+        exponent = math.ceil(number_evaluation_triggers * math.log10(f_evals))
+        value = math.floor(10 ** (exponent / number_evaluation_triggers))
+        return f_evals == value
+
+    @staticmethod
+    def base_eval(
+        base_evaluation_triggers: List[int], dimension: int, f_evals: int
+    ) -> bool:
+        """
+        Determine if the current function evaluation count `f_evals` meets the
+        base evaluation trigger condition.
+
+        - every base_evaluation * dimension * (10**exponent2) for exponent2 >= 0
+        See: https://github.com/numbbo/coco/blob/master/code-experiments/src/coco_observer.c
+        Args:
+            base_evaluation_triggers (List[int]): List of base evaluation triggers.
+            dimension (int): Problem dimension.
+            f_evals (int): Current function evaluation count.
+        Returns:
+            bool: True if the trigger condition is met, False otherwise.
+        """
+
+        if dimension <= 0:
             return False
-        assert (
-            self.best_dist_opt is not None
-        ), "best_dist_opt must be set before checking targets"
-        if self.best_dist_opt <= trigger_targets[0]:
-            # Next target reached.
-            trigger_targets.pop(0)
-            return True
+        for base in base_evaluation_triggers:
+            scaled_eval = f_evals / (dimension * base)
+            assert scaled_eval > 0, "scaled_eval must be positive"
+            # check if scaled_eval is a power of 10
+            if math.log10(scaled_eval).is_integer():
+                return True
         return False
+
+    def emit(  # type: ignore[override]
+        self,
+        base_evaluation_triggers: Optional[List[int]] = None,
+        number_evaluation_triggers: int = 20,
+        last_tdat_emit: Optional[int] = None,  # only pass when emitting the last record
+    ) -> bool:  # type: ignore[override]
+        """
+        Check if the record should be emitted based on the trigger_nth condition.
+        """
+        assert self.f_evals > 0, "f_evals must be positive to determine emission"
+        if base_evaluation_triggers is None:
+            base_evaluation_triggers = [1, 2, 5]
+        if last_tdat_emit is not None:
+            # This is the last evaluation before the end of the run
+            if self.f_evals > last_tdat_emit:
+                # The last evaluation has not already been emitted
+                return True
+            else:  # evaluation already emitted
+                return False
+        if self.f_evals == 1:
+            # Always emit the first evaluation (unless it is also the last)
+            return True
+        trigger_nth = COCOtdatRecord.trigger_nth(
+            number_evaluation_triggers, self.f_evals
+        )
+        trigger_base = COCOtdatRecord.base_eval(
+            base_evaluation_triggers, self.dim, self.f_evals
+        )
+        return trigger_nth or trigger_base
 
 
 class COCOtdatHeader(COCOLogHeader):
